requirements.txt
allure-pytest==2.8.18
pytest==7.2.1
requests==2.28.2
pytest-xdist==2.1.0
selenium==4.5.0
webdriver_manager==2.5.0
pre-commit==2.21.0
black==22.3.0
isort==5.12.0
mypy==0.931
sqlalchemy==1.4.44
psycopg2-binary==2.9.5
keyboard~=0.13.5